+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- **[BREAKING]** support for NodeJS 16
+
 ## [6.3.1] - 2024-08-07
 
 ### Changed